audioflinger: Fix audio for WifiDisplay
AudioFlinger is not able to determine the correct
pid/tid for WifiDisplay and thus we do not pass checks
for CAPTURE_AUDIO_OUTPUT and RECORD_AUDIO permissions.

To fix audio for WifiDisplay, it should be safe to
always allow a trusted calling uid (AID_MEDIA which
has the same perms as AID_AUDIOSERVER).

Change-Id: Ifa46d8e77a43027645cad02a04263b58e134c3ad

 Conflicts:
	services/audiopolicy/service/AudioPolicyInterfaceImpl.cpp

Change-Id: Ieddc8346d0ef6c9b2323f26816f463cba2ab6fc2
[@Jprimero15: adapt after commit: https://github.com/AOSPA/android_frameworks_av/commit/23f552f22abd6bf336484c7085efd46bebd67d89 and added some other changes]
Signed-off-by: Jprimero15 <jprimero155@gmail.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>